+++ lib/rxdart.dart
@@ -1,0 +1,18 @@
+//RxDart 0.26.0
+//https://pub.dev/packages/rxdart/versions/0.26.0
+
+export 'rxdart/behavior_subject.dart';
+export 'rxdart/combine_lastest.dart';
+export 'rxdart/debounce.dart';
+export 'rxdart/defer.dart';
+export 'rxdart/error_and_stacktrace.dart';
+export 'rxdart/forwarding_sink.dart';
+export 'rxdart/forwarding_stream.dart';
+export 'rxdart/merge.dart';
+export 'rxdart/rx.dart';
+export 'rxdart/start_with.dart';
+export 'rxdart/start_with_error.dart';
+export 'rxdart/subject.dart';
+export 'rxdart/switch_map.dart';
+export 'rxdart/value_stream.dart';
+export 'rxdart/value_wrapper.dart';