+++ lib/select_dialog.dart
@@ -11,45 +11,45 @@ typedef Widget ErrorBuilderType<T>(BuildContext context, dynamic exception);
 typedef Widget ButtonBuilderType(BuildContext context, VoidCallback onPressed);
 
 class SelectDialog<T> extends StatefulWidget {
-  final T selectedValue;
-  final List<T> multipleSelectedValues;
-  final List<T> itemsList;
+  final T? selectedValue;
+  final List<T>? multipleSelectedValues;
+  final List<T>? itemsList;
 
   ///![image](https://user-images.githubusercontent.com/16373553/80187339-db365f00-85e5-11ea-81ad-df17d7e7034e.png)
   final bool showSearchBox;
-  final void Function(T) onChange;
-  final void Function(List<T>) onMultipleItemsChange;
-  final Future<List<T>> Function(String text) onFind;
-  final SelectOneItemBuilderType<T> itemBuilder;
-  final WidgetBuilder emptyBuilder;
-  final WidgetBuilder loadingBuilder;
+  final void Function(T)? onChange;
+  final void Function(List<T>)? onMultipleItemsChange;
+  final Future<List<T>> Function(String text)? onFind;
+  final SelectOneItemBuilderType<T>? itemBuilder;
+  final WidgetBuilder? emptyBuilder;
+  final WidgetBuilder? loadingBuilder;
 
   ///![image](https://user-images.githubusercontent.com/16373553/94357272-d599e500-006d-11eb-9bcb-5f067943011e.png)
-  final ButtonBuilderType okButtonBuilder;
-  final ErrorBuilderType errorBuilder;
+  final ButtonBuilderType? okButtonBuilder;
+  final ErrorBuilderType? errorBuilder;
   final bool autofocus;
   final bool alwaysShowScrollBar;
   final int searchBoxMaxLines;
   final int searchBoxMinLines;
 
   ///![image](https://user-images.githubusercontent.com/16373553/80187339-db365f00-85e5-11ea-81ad-df17d7e7034e.png)
-  final InputDecoration searchBoxDecoration;
-
-  final String searchHint;
+  final InputDecoration? searchBoxDecoration;
+  @Deprecated("Use 'hintText' property from searchBoxDecoration")
+  final String? searchHint;
 
   ///![image](https://user-images.githubusercontent.com/16373553/80187103-72e77d80-85e5-11ea-9349-e4dc8ec323bc.png)
-  final TextStyle titleStyle;
+  final TextStyle? titleStyle;
 
   ///|**Max width**: 90% of screen width|**Max height**: 70% of screen height|
   ///|---|---|
   ///|![image](https://user-images.githubusercontent.com/16373553/80189438-0a020480-85e9-11ea-8e63-3fabfa42c1c7.png)|![image](https://user-images.githubusercontent.com/16373553/80190562-e2ac3700-85ea-11ea-82ef-3383ae32ab02.png)|
-  final BoxConstraints constraints;
-  final TextEditingController findController;
+  final BoxConstraints? constraints;
+  final TextEditingController? findController;
 
   const SelectDialog({
-    Key key,
+    Key? key,
     this.itemsList,
-    this.showSearchBox,
+    this.showSearchBox = true,
     this.onChange,
     this.onMultipleItemsChange,
     this.selectedValue,
@@ -57,7 +57,7 @@ class SelectDialog<T> extends StatefulWidget {
     this.onFind,
     this.itemBuilder,
     this.searchBoxDecoration,
-    String searchHint,
+    this.searchHint,
     this.titleStyle,
     this.emptyBuilder,
     this.okButtonBuilder,
@@ -65,51 +65,46 @@ class SelectDialog<T> extends StatefulWidget {
     this.loadingBuilder,
     this.constraints,
     this.autofocus = false,
-    bool alwaysShowScrollBar,
+    this.alwaysShowScrollBar = false,
     this.searchBoxMaxLines = 1,
     this.searchBoxMinLines = 1,
     this.findController,
-  })  : searchHint = searchHint ?? "Find",
-        alwaysShowScrollBar = alwaysShowScrollBar ?? false,
-        super(key: key);
+  }) : super(key: key);
 
-  static Future<T> showModal<T>(
+  static Future<T?> showModal<T>(
     BuildContext context, {
-    List<T> items,
-    String label,
-    T selectedValue,
-    List<T> multipleSelectedValues,
-    bool showSearchBox,
-    Future<List<T>> Function(String text) onFind,
-    SelectOneItemBuilderType<T> itemBuilder,
-    void Function(T) onChange,
-    void Function(List<T>) onMultipleItemsChange,
-    InputDecoration searchBoxDecoration,
-    String searchHint,
-    Color backgroundColor,
-    TextStyle titleStyle,
-    WidgetBuilder emptyBuilder,
-    ButtonBuilderType okButtonBuilder,
-    WidgetBuilder loadingBuilder,
-    ErrorBuilderType errorBuilder,
-    BoxConstraints constraints,
+    List<T>? items,
+    String? label,
+    T? selectedValue,
+    List<T>? multipleSelectedValues,
+    bool showSearchBox = true,
+    Future<List<T>> Function(String text)? onFind,
+    SelectOneItemBuilderType<T>? itemBuilder,
+    void Function(T)? onChange,
+    void Function(List<T>)? onMultipleItemsChange,
+    InputDecoration? searchBoxDecoration,
+    @Deprecated("Use 'hintText' property from searchBoxDecoration") String? searchHint,
+    Color? backgroundColor,
+    TextStyle? titleStyle,
+    WidgetBuilder? emptyBuilder,
+    ButtonBuilderType? okButtonBuilder,
+    WidgetBuilder? loadingBuilder,
+    ErrorBuilderType? errorBuilder,
+    BoxConstraints? constraints,
     bool autofocus = false,
     bool alwaysShowScrollBar = false,
     int searchBoxMaxLines = 1,
     int searchBoxMinLines = 1,
-    TextEditingController findController,
+    TextEditingController? findController,
     bool useRootNavigator = false,
   }) {
-    return showDialog(
+    return showDialog<T>(
       context: context,
       useRootNavigator: useRootNavigator,
       builder: (context) {
         return AlertDialog(
           backgroundColor: backgroundColor,
-          title: Text(
-            label ?? "",
-            style: titleStyle,
-          ),
+          title: Text(label ?? "", style: titleStyle),
           content: SelectDialog<T>(
             selectedValue: selectedValue,
             multipleSelectedValues: multipleSelectedValues,
@@ -150,29 +145,19 @@ class SelectDialog<T> extends StatefulWidget {
 }
 
 class _SelectDialogState<T> extends State<SelectDialog<T>> {
-  SelectOneBloc<T> bloc;
-  MultipleItemsBloc<T> multipleItemsBloc;
-  void Function(T) onChange;
+  late SelectOneBloc<T> bloc;
+  late MultipleItemsBloc<T> multipleItemsBloc;
+  void Function(T)? onChange;
 
-  _SelectDialogState(
-    List<T> itemsList,
-    this.onChange,
-    void Function(List<T>) onMultipleItemsChange,
-    List<T> multipleSelectedValues,
-    Future<List<T>> Function(String text) onFind,
-    TextEditingController findController,
-  ) {
+  _SelectDialogState(List<T>? itemsList, this.onChange, void Function(List<T>)? onMultipleItemsChange, List<T>? multipleSelectedValues, Future<List<T>> Function(String text)? onFind, TextEditingController? findController) {
     bloc = SelectOneBloc(itemsList, onFind, findController);
-    multipleItemsBloc = MultipleItemsBloc(
-      multipleSelectedValues,
-      onMultipleItemsChange,
-    );
+    multipleItemsBloc = MultipleItemsBloc(multipleSelectedValues, onMultipleItemsChange);
   }
 
   @override
   void didChangeDependencies() {
     super.didChangeDependencies();
-    if (widget.autofocus ?? false) {
+    if (widget.autofocus) {
       FocusScope.of(context).requestFocus(bloc.focusNode);
     }
   }
@@ -194,21 +179,9 @@ class _SelectDialogState<T> extends State<SelectDialog<T>> {
         maxHeight: MediaQuery.of(context).size.height * 0.7,
       );
 
-  SelectOneItemBuilderType<T> get itemBuilder =>
-      widget.itemBuilder ??
-      (context, item, isSelected) => ListTile(
-            title: Text(item.toString()),
-            selected: isSelected,
-          );
+  SelectOneItemBuilderType<T> get itemBuilder => widget.itemBuilder ?? (context, item, isSelected) => ListTile(title: Text(item.toString()), selected: isSelected);
 
-  ButtonBuilderType get okButtonBuilder =>
-      widget.okButtonBuilder ??
-      (context, onPressed) {
-        return RaisedButton(
-          child: Text("Ok"),
-          onPressed: onPressed,
-        );
-      };
+  ButtonBuilderType get okButtonBuilder => widget.okButtonBuilder ?? (context, onPressed) => ElevatedButton(child: Text("Ok"), onPressed: onPressed);
 
   @override
   Widget build(BuildContext context) {
@@ -218,7 +191,7 @@ class _SelectDialogState<T> extends State<SelectDialog<T>> {
       constraints: widget.constraints ?? (isWeb ? webDefaultConstraints : mobileDefaultConstraints),
       child: Column(
         children: <Widget>[
-          if (widget.showSearchBox ?? true)
+          if (widget.showSearchBox)
             Padding(
               padding: const EdgeInsets.all(8.0),
               child: TextField(
@@ -226,57 +199,35 @@ class _SelectDialogState<T> extends State<SelectDialog<T>> {
                 focusNode: bloc.focusNode,
                 maxLines: widget.searchBoxMaxLines,
                 minLines: widget.searchBoxMinLines,
-                decoration: widget.searchBoxDecoration ??
-                    InputDecoration(
-                      hintText: widget.searchHint,
-                      contentPadding: const EdgeInsets.all(2.0),
-                    ),
+                decoration: widget.searchBoxDecoration?.copyWith(hintText: widget.searchHint ?? widget.searchBoxDecoration!.hintText) ?? InputDecoration(hintText: widget.searchHint ?? "Find", contentPadding: const EdgeInsets.all(2.0)),
               ),
             ),
           Expanded(
-            child: StreamBuilder<List<T>>(
+            child: StreamBuilder<List<T>?>(
               stream: bloc.filteredListOut,
               builder: (context, snapshot) {
                 if (snapshot.hasError) {
-                  if (widget.errorBuilder != null) {
-                    return widget.errorBuilder(context, snapshot.error);
-                  } else {
-                    return Center(child: Text("Oops. \n${snapshot.error}"));
-                  }
+                  return widget.errorBuilder?.call(context, snapshot.error) ?? Center(child: Text("Oops. \n${snapshot.error}"));
                 } else if (!snapshot.hasData) {
-                  if (widget.loadingBuilder != null) {
-                    return widget.loadingBuilder(context);
-                  } else {
-                    return Center(child: CircularProgressIndicator());
-                  }
-                } else if (snapshot.data.isEmpty) {
-                  if (widget.emptyBuilder != null) {
-                    return widget.emptyBuilder(context);
-                  } else {
-                    return Center(child: Text("No data found"));
-                  }
+                  return widget.loadingBuilder?.call(context) ?? Center(child: CircularProgressIndicator());
+                } else if (snapshot.data!.isEmpty) {
+                  return widget.emptyBuilder?.call(context) ?? Center(child: Text("No data found"));
                 }
                 return Scrollbar(
                   controller: bloc.scrollController,
                   isAlwaysShown: widget.alwaysShowScrollBar,
                   child: ListView.builder(
                     controller: bloc.scrollController,
-                    itemCount: snapshot.data.length,
+                    itemCount: snapshot.data!.length,
                     itemBuilder: (context, index) {
-                      var item = snapshot.data[index];
-                      bool isSelected = multipleItemsBloc.selectedItems?.contains(item) ?? false;
+                      var item = snapshot.data![index];
+                      bool isSelected = multipleItemsBloc.selectedItems.contains(item);
                       isSelected = isSelected || item == widget.selectedValue;
                       return InkWell(
                         child: itemBuilder(context, item, isSelected),
                         onTap: () {
                           if (isMultipleItems) {
-                            setState(() {
-                              if (isSelected) {
-                                multipleItemsBloc.unselectItem(item);
-                              } else {
-                                multipleItemsBloc.selectItem(item);
-                              }
-                            });
+                            setState(() => (isSelected) ? multipleItemsBloc.unselectItem(item) : multipleItemsBloc.selectItem(item));
                           } else {
                             onChange?.call(item);
                             Navigator.pop(context);